+++ phase2/code/src/phase2_baselines/runners/__init__.py
@@ -2,5 +2,6 @@
 
 from .react import ReactRunner
 from .single_path import SinglePathRunner
+from .tot import ToTRunner
 
-__all__ = ["SinglePathRunner", "ReactRunner"]
+__all__ = ["SinglePathRunner", "ReactRunner", "ToTRunner"]